api/requirements.txt
fastapi[standard]
PyJWT
google-auth
requests
boto3